fix: ass rootDir in tsconfig.json
